[refactor] dom을 사용하는 box 추상화
