+++ src/phaser/scene/ShopManagerScene/uiObjects/DomBox/index.js
@@ -1,0 +1,45 @@
+import Phaser from '@/package/phaser';
+
+class TextBox extends Phaser.GameObjects.DOMElement {
+  constructor(scene, domWrapperId, width = 0, height = 0) {
+    super(scene).createFromHTML(
+      `<div
+        id="${domWrapperId}"
+        style="width: ${width}px; height: ${height}px;"
+      ></div>
+      <style type="text/css">
+        #${domWrapperId}::-webkit-scrollbar {
+          display: none;
+        }
+
+        #${domWrapperId} {
+          overflow-y: scroll;
+        }
+      </style>`,
+    );
+
+    this._domWrapper = this.getChildByID(`${domWrapperId}`);
+
+    scene.add.existing(this);
+  }
+
+  _clearHTML() {
+    while (this._domWrapper.firstChild) this._domWrapper.removeChild(this._domWrapper.lastChild);
+  }
+
+  _setHTML(html) {
+    this._clearHTML();
+    this._domWrapper.insertAdjacentHTML('beforeend', html);
+  }
+
+  _setColor(color) {
+    this._domWrapper.style.background = `#${color.toString(16)}`;
+  }
+
+  setSize(width, height) {
+    this._domWrapper.style.width = `${width}px`;
+    this._domWrapper.style.height = `${height}px`;
+  }
+}
+
+export default TextBox;
